refactor: mv all classify fns to utils
